Refactor JSON formatters
