+++ lib/ougai/formatters/bunyan.rb
@@ -1,5 +1,4 @@
 require 'ougai/formatters/base'
-require 'oj'
 
 module Ougai
   module Formatters
@@ -38,17 +37,8 @@ module Ougai
         }.merge(data))
       end
 
-      private
-
-      OJ_OPTIONS = { mode: :custom, time_format: :xmlschema,
-                     use_as_json: true, use_to_hash: true, use_to_json: true }
-
-      def dump(data)
-        return data unless @jsonize
+      def convert_time(data)
         data[:time] = format_datetime(data[:time])
-        str = Oj.dump(data, OJ_OPTIONS)
-        str << "\n" if @with_newline
-        str
       end
     end
   end
