+++ lib/ougai/formatters/for_json.rb
@@ -1,3 +1,5 @@
+require 'oj'
+
 module Ougai
   # The features for JSON
   module Formatters::ForJson
@@ -18,6 +20,18 @@ module Ougai
       else
         70
       end
+    end
+
+    OJ_OPTIONS = { mode: :custom, time_format: :xmlschema,
+                   use_as_json: true, use_to_hash: true, use_to_json: true }
+
+    # requires convert_time(data) method
+    def dump(data)
+      return data unless @jsonize
+      convert_time(data)
+      str = Oj.dump(data, OJ_OPTIONS)
+      str << "\n" if @with_newline
+      str
     end
   end
 end
